Add .gitignore file to exclude unnecessary files
This commit adds a .gitignore file to the repository to specify which files and directories should be ignored by Git. 
The .gitignore file includes patterns for temporary files, environment files, and other artifacts that do not need to be tracked in version control.
This ensures a cleaner repository by preventing unnecessary files from being included in commits and maintaining focus on the relevant project files.

Signed-off-by: Jeevitha Shree.S <jeevishree4@gmail.com>